+++ api/api.iml
@@ -7,11 +7,8 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="module" module-name="common" exported="" />
-    <orderEntry type="module" module-name="commonView" exported="" />
-    <orderEntry type="module" module-name="api" exported="" />
-    <orderEntry type="module" module-name="merging" exported="" />
-    <orderEntry type="module" module-name="shellSort" exported="" />
-    <orderEntry type="module" module-name="simpleSort" exported="" />
+    <orderEntry type="module" module-name="simpleSort" />
+    <orderEntry type="module" module-name="merging" />
+    <orderEntry type="module" module-name="shellSort" />
   </component>
 </module>
