+++ api/src/ApiUtils/SortApiUtil.java
@@ -1,0 +1,40 @@
+package ApiUtils;
+
+
+import sort.Merg.Merg;
+import sort.Merg.impl.MergImpl;
+import sort.defaultSort.Bubble;
+import sort.defaultSort.Insertion;
+import sort.defaultSort.Selection;
+import sort.defaultSort.impl.BubbleImpl;
+import sort.defaultSort.impl.InsertionImpl;
+import sort.defaultSort.impl.SelectionImpl;
+import sort.shellSort.Shell;
+import sort.shellSort.impl.ShellImpl;
+
+public class SortApiUtil {
+    static Bubble BUBBLE = new BubbleImpl();
+    public static Bubble BUBBLE(Bubble bubble){
+        return bubble;
+    }
+
+    static Selection SELECTION= new SelectionImpl();
+    public static Selection SELECTION(Selection selection){
+        return selection;
+    }
+
+    static Insertion INSERTION = new InsertionImpl();
+    public static Insertion INSERTION(Insertion insertion){
+        return insertion;
+    }
+
+    static Shell SHELL = new ShellImpl();
+    public static Shell SHELL(Shell shell){
+        return shell;
+    }
+
+    static Merg MERG = new MergImpl();
+    public static Merg MERG(Merg merg) {
+        return merg;
+    }
+}
