+++ commonView/src/visual/view/ViewDataHelper.java
@@ -1,4 +1,4 @@
-package view;
+package visual.view;
 
 import java.awt.*;
 import java.awt.geom.Ellipse2D;
