+++ visualization/src/visual/model/SimpleSortData.java
@@ -1,4 +1,4 @@
-package model;
+package visual.model;
 
 public class SimpleSortData extends SortData {
